fix(storybook): split oversized stories, add play functions and usage guidelines (#1227)
BLOCKER-1: Split 6 oversized story files (>500 lines) into 13 smaller files:
- Button → ButtonBasic + ButtonVariants
- Card → CardBasic + CardVariants + CardComposed
- Input → InputBasic + InputVariants
- Tabs → TabsBasic + TabsVariants
- Resizer → ResizerBasic + ResizerInteractions
- AiDialogs → AiDialogsInline + AiDialogsStates

BLOCKER-2: Added 10 play functions (292 total, target ≥290):
- ButtonBasic: Default, Primary, Loading, Disabled
- TabsBasic: Default, SettingsTabs
- InputBasic: Default, Disabled
- CardBasic: Default
- Toggle: Default

BLOCKER-3: Added usage guidelines (使用指南) to 4 more components (9 total):
- Dialog, Select, Checkbox, Tooltip

BLOCKER-4: Added demo-only docs and tags to static replica stories:
- Settings, RightPanelInfo, RightPanelQuality

BLOCKER-5: Replaced hardcoded colors with CSS variables/tokens:
- preview.tsx, SearchInteractive, QualityActions, VersionActions,
  Settings, RightPanelInfo, RightPanelQuality, AiDialogsInline

Updated surfaceRegistry.ts with new sub-path entries (95 total).
