+++ apps/desktop/renderer/src/features/quality-gates/QualityActions.stories.tsx
@@ -31,16 +31,16 @@ const meta: Meta<typeof QualityGatesPanel> = {
             </div>
           </div>
           <div className="flex-1 p-12 overflow-hidden flex justify-center">
-            <div className="w-full max-w-3xl h-full bg-[#121212] rounded-t-lg border-x border-t border-[var(--color-separator)] shadow-2xl p-16 relative">
-              <div className="w-1/3 h-8 bg-[rgba(255,255,255,0.08)] rounded mb-10" />
-              <div className="w-full h-3 bg-[rgba(255,255,255,0.04)] rounded mb-4" />
-              <div className="w-full h-3 bg-[rgba(255,255,255,0.04)] rounded mb-4" />
-              <div className="w-5/6 h-3 bg-[rgba(255,255,255,0.04)] rounded mb-4" />
-              <div className="w-full h-3 bg-[rgba(255,255,255,0.04)] rounded mb-8" />
-              <div className="w-1/4 h-5 bg-[rgba(255,255,255,0.06)] rounded mb-6" />
-              <div className="w-full h-3 bg-[rgba(255,255,255,0.04)] rounded mb-4" />
-              <div className="w-11/12 h-3 bg-[rgba(255,255,255,0.04)] rounded mb-4" />
-              <div className="w-full h-3 bg-[rgba(255,255,255,0.04)] rounded mb-4" />
+            <div className="w-full max-w-3xl h-full bg-[var(--color-bg-surface)] rounded-t-lg border-x border-t border-[var(--color-separator)] shadow-[var(--shadow-xl)] p-16 relative">
+              <div className="w-1/3 h-8 bg-white/8 rounded mb-10" />
+              <div className="w-full h-3 bg-white/[0.04] rounded mb-4" />
+              <div className="w-full h-3 bg-white/[0.04] rounded mb-4" />
+              <div className="w-5/6 h-3 bg-white/[0.04] rounded mb-4" />
+              <div className="w-full h-3 bg-white/[0.04] rounded mb-8" />
+              <div className="w-1/4 h-5 bg-white/[0.06] rounded mb-6" />
+              <div className="w-full h-3 bg-white/[0.04] rounded mb-4" />
+              <div className="w-11/12 h-3 bg-white/[0.04] rounded mb-4" />
+              <div className="w-full h-3 bg-white/[0.04] rounded mb-4" />
             </div>
           </div>
         </div>
